Changed posts processing to API section and image styling
